build(deps): upgrade vitest to v5
vitest 4.0.18 -> 5.0.0 (plus @vitest/coverage-istanbul and @vitest/ui),
vite 6 -> 8.3.0, jsdom 29 -> 30, and @vitejs/plugin-react 4 -> 6 in the
two examples that use it. vitest 5 requires Node 22.12+ and Vite >=6.4.

Three things had to change:

- The root config is split into `vitest.shared.ts` (all the shared test
  options) and `vitest.config.ts` (shared + `projects`). vitest 5 treats
  any config declaring `projects` as a container, and every package's
  config re-exported the root one - so each package became a container
  looking for `./packages/*` relative to itself and failed with "No
  projects were found in packages/client/vitest.config.ts".

- `packages/upgrade` sets `jsx: preserve` in its tsconfig, which is right
  for a codemod package: the fixtures have to keep their JSX intact for
  jscodeshift. Vite 8 honours that setting where Vite 6 did not, so
  fixture JSX reached the SSR transform uncompiled and every `hooks`
  transform test died with "Unexpected JSX expression". Its vitest config
  now overrides the JSX runtime, since the tests need the fixtures
  actually executed.

- `transforms.test.ts` loads fixtures through `import.meta.glob` instead
  of a dynamic `import()` of an absolute path. Vite only transforms
  modules it knows about, and a path computed at runtime is not one. It
  also calls `cleanup()` after each spec run: every fixture renders
  inside a single `it`, so nothing unmounted until the test ended, and a
  React commit firing on a timer after jsdom teardown threw
  "Right-hand side of 'instanceof' is not an object". vitest 5 fails a
  run on unhandled errors where 4 did not.

Vite 8's native config loader warnings are dealt with too: the two vitest
configs living in CommonJS packages are renamed to `.mts`, and the shared
config is imported with its real `.ts` extension (hence
`allowImportingTsExtensions` in the typecheck tsconfig). That rename also
needs `.ts-prunerc` to ignore `config.mts`, not just `config.ts`, or
`lint-prune` flags the re-export and fails autofix.ci.

`examples/minimal-react/client` and `examples/minimal-content-types/client`
move to `moduleResolution: Bundler`. Vite 8 ships its types only through
`exports`, which node10 resolution cannot read, so `tsc` could no longer
find `vite` or `@vitejs/plugin-react` at all.

`nodeModulesJsDocRouter`'s committed fixtures are regenerated. That
router deliberately derives its schema from types in node_modules, and
rolldown's `buildStart` hook type changed with the Vite bump - so the
generated OpenAPI document legitimately changes with it.
