+++ examples/minimal-react/client/package.json
@@ -19,9 +19,9 @@
   "devDependencies": {
     "@types/react": "^19.1.0",
     "@types/react-dom": "^19.1.1",
-    "@vitejs/plugin-react": "^4.3.4",
+    "@vitejs/plugin-react": "^6.1.1",
     "eslint": "^10.10.0",
     "typescript": "^5.9.2",
-    "vite": "^6.1.1"
+    "vite": "^8.3.0"
   }
 }
