+++ examples/next-prisma-starter/package.json
@@ -64,8 +64,8 @@
     "tsx": "^4.19.3",
     "typescript": "^5.9.2",
     "typescript-eslint": "^8.70.0",
-    "vite": "^6.1.1",
-    "vitest": "^4.0.18"
+    "vite": "^8.3.0",
+    "vitest": "^5.0.0"
   },
   "publishConfig": {
     "access": "restricted"
