+++ packages/openapi/vitest.config.ts
@@ -1,5 +1,5 @@
 import { defineConfig, mergeConfig } from 'vitest/config';
-import base from '../../vitest.config';
+import base from '../../vitest.shared.ts';
 
 export default mergeConfig(
   base,
