+++ packages/react-query/vitest.config.ts
@@ -1,5 +1,5 @@
 import { mergeConfig } from 'vitest/config';
-import baseConfig from '../../vitest.config';
+import baseConfig from '../../vitest.shared.ts';
 
 export default mergeConfig(baseConfig, {
   test: {
