+++ packages/tests/package.json
@@ -21,7 +21,7 @@
     "eslint": "catalog:",
     "fastify": "^5.0.0",
     "fastify-plugin": "^5.0.0",
-    "jsdom": "^29.0.0",
+    "jsdom": "catalog:",
     "myzod": "^1.3.1",
     "node-fetch": "^3.3.0",
     "runtypes": "^7.0.0",
