+++ packages/tests/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "../../tsconfig.json",
-  "include": ["server", "showcase", "vitest.config.ts"],
+  "include": ["server", "showcase", "vitest.config.mts"],
   "compilerOptions": {
     "skipLibCheck": true,
     "lib": ["ES2023", "DOM", "DOM.Iterable"],
