--- packages/tests/vitest.config.ts
@@ -1,1 +1,0 @@
-export { default } from '../../vitest.config';
