+++ pnpm-workspace.yaml
@@ -54,6 +54,11 @@ catalog:
   prettier: ^3.9.6
   prettier-plugin-tailwindcss: ^0.8.1
   '@ianvs/prettier-plugin-sort-imports': ^4.7.1
+  vitest: ^5.0.0
+  '@vitest/coverage-istanbul': ^5.0.0
+  '@vitest/ui': ^5.0.0
+  vite: ^8.3.0
+  jsdom: ^30.0.1
 
 catalogs:
   ts-api:
